SikeResume-main/requirements.txt
langchain==0.2.12
streamlit==1.37.0
pypdf2==3.0.1
python-dotenv==1.0.1
langchain-community==0.2.11
google-generativeai==0.7.2
faiss-cpu
langchain-google-genai==1.0.8
langsmith==0.1.106